+++ mex/datscha_web/main.py
@@ -25,8 +25,8 @@ from mex.datscha_web.transform import transform_datscha_web_items_to_mex_activit
 from mex.filters import filter_by_global_rules
 from mex.pipeline import asset, run_job_in_process
 from mex.sinks import load
-from mex.wikidata.transform import (
-    transform_wikidata_organizations_to_extracted_organizations_with_query,
+from mex.wikidata.extract import (
+    get_merged_organization_id_by_query_with_transform_and_load,
 )
 
 
@@ -87,17 +87,9 @@ def datscha_web_organization_ids_by_query_string(
         extracted_datscha_web_items
     )
 
-    extracted_organizations_by_query = (
-        transform_wikidata_organizations_to_extracted_organizations_with_query(
-            wikidata_organizations_by_query, extracted_primary_source_wikidata
-        )
+    return get_merged_organization_id_by_query_with_transform_and_load(
+        wikidata_organizations_by_query, extracted_primary_source_wikidata
     )
-    load(extracted_organizations_by_query.values())
-
-    return {
-        query: MergedOrganizationIdentifier(organization.stableTargetId)
-        for query, organization in extracted_organizations_by_query.items()
-    }
 
 
 @asset(group_name="datscha_web")