+++ mex/international_projects/main.py
@@ -8,15 +8,11 @@ from mex.common.types import (
     MergedOrganizationIdentifier,
     MergedPersonIdentifier,
 )
-from mex.common.wikidata.transform import (
-    transform_wikidata_organizations_to_extracted_organizations,
-)
 from mex.international_projects.extract import (
     extract_international_projects_funding_sources,
     extract_international_projects_partner_organizations,
     extract_international_projects_project_leaders,
     extract_international_projects_sources,
-    get_organization_merged_id_by_query,
 )
 from mex.international_projects.models.source import InternationalProjectsSource
 from mex.international_projects.settings import InternationalProjectsSettings
@@ -25,6 +21,9 @@ from mex.international_projects.transform import (
 )
 from mex.pipeline import asset, run_job_in_process
 from mex.sinks import load
+from mex.wikidata.extract import (
+    get_merged_organization_id_by_query_with_transform_and_load,
+)
 
 
 @asset(group_name="international_projects", deps=["extracted_primary_source_mex"])
@@ -75,15 +74,8 @@ def international_projects_funding_sources_ids_by_query(
     wikidata_funding_sources_by_query = extract_international_projects_funding_sources(
         international_projects_sources
     )
-    mex_extracted_organizations_funding_sources = (
-        transform_wikidata_organizations_to_extracted_organizations(
-            wikidata_funding_sources_by_query.values(),
-            extracted_primary_source_wikidata,
-        )
-    )
-    load(mex_extracted_organizations_funding_sources)
 
-    return get_organization_merged_id_by_query(
+    return get_merged_organization_id_by_query_with_transform_and_load(
         wikidata_funding_sources_by_query, extracted_primary_source_wikidata
     )
 
@@ -99,15 +91,8 @@ def international_projects_partner_organization_ids_by_query(
             international_projects_sources
         )
     )
-    mex_extracted_organizations_partner_organizations = (
-        transform_wikidata_organizations_to_extracted_organizations(
-            wikidata_partner_organizations_by_query.values(),
-            extracted_primary_source_wikidata,
-        )
-    )
-    load(mex_extracted_organizations_partner_organizations)
 
-    return get_organization_merged_id_by_query(
+    return get_merged_organization_id_by_query_with_transform_and_load(
         wikidata_partner_organizations_by_query, extracted_primary_source_wikidata
     )