+++ mex/odk/main.py
@@ -12,14 +12,10 @@ from mex.common.types import (
     MergedOrganizationalUnitIdentifier,
     MergedOrganizationIdentifier,
 )
-from mex.common.wikidata.transform import (
-    transform_wikidata_organizations_to_extracted_organizations,
-)
 from mex.mapping.extract import extract_mapping_data
 from mex.odk.extract import (
     extract_odk_raw_data,
     get_external_partner_and_publisher_by_label,
-    get_organization_merged_id_by_query,
 )
 from mex.odk.model import ODKData
 from mex.odk.settings import ODKSettings
@@ -31,6 +27,9 @@ from mex.odk.transform import (
 )
 from mex.pipeline import asset, run_job_in_process
 from mex.sinks import load
+from mex.wikidata.extract import (
+    get_merged_organization_id_by_query_with_transform_and_load,
+)
 
 
 @asset(group_name="odk", deps=["extracted_primary_source_mex"])
@@ -71,15 +70,7 @@ def external_partner_and_publisher_by_label(
         get_external_partner_and_publisher_by_label(odk_resource_mappings)
     )
 
-    mex_extracted_organizations_partner_organizations = (
-        transform_wikidata_organizations_to_extracted_organizations(
-            wikidata_partner_organizations_by_query.values(),
-            extracted_primary_source_wikidata,
-        )
-    )
-    load(mex_extracted_organizations_partner_organizations)
-
-    return get_organization_merged_id_by_query(
+    return get_merged_organization_id_by_query_with_transform_and_load(
         wikidata_partner_organizations_by_query, extracted_primary_source_wikidata
     )