+++ mex/sumo/main.py
@@ -22,9 +22,7 @@ from mex.common.types import (
     Email,
     MergedContactPointIdentifier,
     MergedOrganizationalUnitIdentifier,
-)
-from mex.common.wikidata.transform import (
-    transform_wikidata_organizations_to_extracted_organizations,
+    MergedOrganizationIdentifier,
 )
 from mex.mapping.extract import extract_mapping_data
 from mex.pipeline import asset, run_job_in_process
@@ -39,7 +37,6 @@ from mex.sumo.extract import (
     extract_ldap_contact_points_by_emails,
     extract_ldap_contact_points_by_name,
     extract_sumo_organizations,
-    get_organization_merged_id_by_query,
 )
 from mex.sumo.filter import filter_and_log_cc2_aux_model
 from mex.sumo.models.cc1_data_model_nokeda import Cc1DataModelNoKeda
@@ -59,6 +56,9 @@ from mex.sumo.transform import (
     transform_resource_nokeda_to_mex_resource,
     transform_sumo_access_platform_to_mex_access_platform,
     transform_sumo_activity_to_extracted_activity,
+)
+from mex.wikidata.extract import (
+    get_merged_organization_id_by_query_with_transform_and_load,
 )
 
 
@@ -199,16 +199,11 @@ def extracted_cc2_feat_projection() -> list[Cc2FeatProjection]:
 def organization_stable_target_id_by_query_sumo(
     extracted_resources_nokeda_sumo: dict[str, Any],
     extracted_primary_source_wikidata: ExtractedPrimarySource,
-) -> dict[str, str]:
+) -> dict[str, MergedOrganizationIdentifier]:
     """Extract and load SUMO organizations and return them by stable target ids."""
     sumo_organizations = extract_sumo_organizations(extracted_resources_nokeda_sumo)
-    mex_extracted_organizations = list(
-        transform_wikidata_organizations_to_extracted_organizations(
-            sumo_organizations.values(), extracted_primary_source_wikidata
-        )
-    )
-    load(mex_extracted_organizations)
-    return get_organization_merged_id_by_query(
+
+    return get_merged_organization_id_by_query_with_transform_and_load(
         sumo_organizations, extracted_primary_source_wikidata
     )